pyproject.toml
[project]
name = "gemma-forge"
version = "0.0.1"
description = "Air-gapped, agentic infrastructure remediation on the Dell PowerEdge XR7620. Gemma 4 + vLLM + Google ADK Ralph loops."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Ken Rollins" }]
keywords = ["llm", "agents", "vllm", "gemma", "edge-ai", "federal", "stig", "ralph-loop"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    # Agent framework — pre-1.0, pin tightly and re-check release notes per phase
    "google-adk>=0.3.0,<0.5.0",

    # vLLM client (OpenAI-compatible REST via httpx; we deliberately do NOT use LiteLLM)
    "httpx[http2]>=0.27,<1.0",
    "openai>=1.40,<2.0",

    # Config + schemas
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "pyyaml>=6.0,<7.0",

    # SSH to the target VM(s)
    "asyncssh>=2.14,<3.0",

    # Observability — OTel-pure stack (Jaeger + Prometheus + Grafana).
    # NO Langfuse: see docs/adr/0007-otel-jaeger-prometheus-grafana.md.
    # Token accounting flows through the OpenTelemetry GenAI semantic
    # conventions (gen_ai.* attributes).
    "opentelemetry-api>=1.25,<2.0",
    "opentelemetry-sdk>=1.25,<2.0",
    "opentelemetry-exporter-otlp>=1.25,<2.0",
    "opentelemetry-instrumentation-httpx>=0.46b0",
    "opentelemetry-semantic-conventions>=0.46b0",

    # NVML for GPU telemetry surfaced to the dashboard
    "nvidia-ml-py>=12.0,<13.0",

    # FastAPI surface for the web frontend
    "fastapi>=0.111,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "websockets>=12.0,<14.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5,<1.0",
    "mypy>=1.10,<2.0",
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=5.0,<6.0",
    "types-PyYAML",
]

# Host-operator tooling. Install with `pip install -e ".[infra]"` (or
# `[dev,infra]` for everything). Deliberately separate from `dev` and
# from runtime `dependencies` so the harness Docker image (Phase 6+)
# does NOT carry the model-download tooling. These deps are only
# needed on the box that downloads weights into /data/triton/weights/
# during Phase 1 host setup.
infra = [
    "huggingface_hub>=0.26,<1.0",
    "hf-transfer>=0.1.8,<1.0",  # accelerated parallel downloads for HF
]

[project.urls]
Homepage = "https://github.com/kenrollins/gemma-forge"
Repository = "https://github.com/kenrollins/gemma-forge"
Issues = "https://github.com/kenrollins/gemma-forge/issues"

[project.scripts]
gemma-forge = "gemma_forge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gemma_forge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: integration tests requiring real LLM + VM; use --run-slow to include",
]
